Refactor response objects
