+++ src/SuccessResponse.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Drip;
+
+/**
+ * A successful response
+ */
+class SuccessResponse extends AbstractResponse {
+}
